+++ src/core/shared/enums/webViewEvents.ts
@@ -7,4 +7,5 @@ export const WebViewEventNames = {
     SET_PAGES: 'athena:view:set:pages',
     CLOSE_PAGE: 'athena:Close:Page',
     PLAY_SOUND: 'athena:Play:Sound',
+    PLAY_SOUND_FRONTEND: 'athena:Play:Sound:Frontend',
 };